MIM-159 split Browser runtime evidence
